requirements.txt
# ── Core pipeline ─────────────────────────────────────────────────────────────
dbt-core==1.8.0
dbt-duckdb==1.8.0
duckdb==0.10.0
pandas==2.2.0
faker==24.0.0

# ── Kafka streaming ───────────────────────────────────────────────────────────
kafka-python==2.0.2

# ── Airflow ───────────────────────────────────────────────────────────────────
# Install separately to avoid conflicts with dbt dependencies:
#   pip install apache-airflow==2.9.1
# Or use Docker (recommended): docker compose up -d
